Rename to WooPayments not WCPay for new classes
